+++ javascript/WindowsLepik.d.ts
@@ -193,6 +193,7 @@ declare class WindowsLepik extends LepikEvents {
         height: number;
     }>;
     setWindowSize(windowHandle: string, width: number, height: number): void;
+    setWindowPosition(windowHandle: string, x: number, y: number): void;
     /**
      * Delays the execution for the specified number of milliseconds.
      * @param {number} ms - The number of milliseconds to delay.